+++ src/structures.h
@@ -2,14 +2,14 @@
 #define STRUCTURES_H
 
 #if !defined(__HIPSYCL__) && !defined(__CUDACC__) && !defined(__HIPCC__)
-  struct double3 { double x, y, z; };
+  struct float3 { double x, y, z; };
 #else
   // nvcc will pull in '/usr/include/vector_types.h' which 
-  // contains an identical definition of 'double3' (above).
+  // contains an identical definition of 'float3' (above).
 #endif
 
 #define DEBUGGABLE_ABORT fprintf(stderr, "%s:%d\n", __FILE__, __LINE__); fflush(stderr); fflush(stdout); exit(EXIT_FAILURE);
-struct edge { double a, b; };
-struct edge_neighbour { int a, b; double x, y, z; };
+struct edge { float a, b; };
+struct edge_neighbour { int a, b; float x, y, z; };
 
 #endif
